Implement not and neq (#40)
* Implement and test not and != operator. Fixes #38

* Implement (NOT) NULL. Fixes #37

* Stylefix

* Thanks to @kyprifog and @martindurant, fixed the IS NULL operator for
non-numeric types.

+++ tests/integration/test_function.py
@@ -10,7 +10,7 @@ class FunctionTestCase(DaskTestCase):
         def f(x):
             return x ** 2
 
-        self.c.register_function(f, "f", [("x", np.int64)], np.int64)
+        self.c.register_function(f, "f", [("x", np.float64)], np.float64)
 
         df = self.c.sql(
             """
